gcc: fix signness warnings with explicit casts
Add explicit type casts to resolve GCC signness comparison warnings:
- Cast dest_size to int in lvmlockctl.c
- Cast NVME_IDENTIFY_DATA_SIZE to int in nvme.c
- Cast bitwise AND results to int in nvme.c and persist.c
- Cast DM_STATS_GROUP_NOT_PRESENT to int in libdm-stats.c

+++ daemons/lvmlockd/lvmlockctl.c
@@ -164,7 +164,7 @@ static int szscanf(const char *input, const char *format, ...)
 
 				/* copy non-whitespace characters from input to dest */
 				n = 0;
-				while (*in != '\0' && !isspace((unsigned char)*in) && (n < dest_size-1)) {
+				while (*in != '\0' && !isspace((unsigned char)*in) && (n < (int)dest_size-1)) {
 					*out = *in;
 					out++;
 					in++;
